cpu: rv64: add JIT f16 support for binary and eltwise to existing files.
