+++ src/cpu/rv64/jit_rvv_binary_kernel.hpp
@@ -1,5 +1,6 @@
 /*******************************************************************************
 * Copyright 2026 Institute of Software, Chinese Academy of Sciences
+* Copyright 2026 openKylin community
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
@@ -57,6 +58,40 @@ bool jit_rvv_binary_f32_supported(alg_kind_t alg);
 
 void jit_rvv_binary_apply_f32(alg_kind_t alg, const float *src0,
         const float *src1, const int8_t *src2, float *dst, dim_t len);
+
+// f16 binary kernel (native Zvfh arithmetic).
+struct jit_rvv_binary_kernel_f16_t : public jit_generator_t {
+    struct call_params_t {
+        const void *src0;
+        const void *src1;
+        const int8_t *src2;
+        void *dst;
+        dim_t len;
+    };
+
+    DECLARE_CPU_JIT_AUX_FUNCTIONS(jit_rvv_binary_kernel_f16_t)
+
+    explicit jit_rvv_binary_kernel_f16_t(alg_kind_t alg);
+
+    void operator()(const call_params_t *p) const {
+        jit_generator_t::operator()(p);
+    }
+
+protected:
+    void generate() override;
+
+private:
+    void compute_vector(const Xbyak_riscv::VReg &v_dst,
+            const Xbyak_riscv::VReg &v_src0, const Xbyak_riscv::VReg &v_src1,
+            const Xbyak_riscv::FReg &f_zero, const Xbyak_riscv::FReg &f_one);
+
+    alg_kind_t alg_;
+};
+
+bool jit_rvv_binary_f16_supported(alg_kind_t alg);
+
+void jit_rvv_binary_apply_f16(alg_kind_t alg, const void *src0,
+        const void *src1, const int8_t *src2, void *dst, dim_t len);
 
 } // namespace rv64
 } // namespace cpu
